moved load_netcdf4 and computational_form to core
