+++ pysat/utils/__init__.py
@@ -9,5 +9,5 @@ of formats, loading of files, and user-supplied info
 for the pysat data directory structure.
 """
 
-from . import coords, misc, stats, time
-from ._core import set_data_dir
+from . import coords, stats, time
+from ._core import set_data_dir, computational_form, load_netcdf4
